Removed confusing BileMo namespace before AppBundle
