+++ app/config/security.yml
@@ -3,7 +3,7 @@
 security:
 
     encoders:
-        BileMo\AppBundle\Entity\User: bcrypt
+        AppBundle\Entity\User: bcrypt
 
     # https://symfony.com/doc/current/security.html#b-configuring-how-users-are-loaded
     providers:
