+++ composer.json
@@ -4,8 +4,7 @@
     "type": "project",
     "autoload": {
         "psr-4": {
-            "BileMo\\AppBundle\\": "src/BileMo/AppBundle",
-            "OAuthServerBundle\\": "src/OAuthServerBundle"
+            "AppBundle\\": "src/AppBundle"
         },
         "classmap": [
             "app/AppKernel.php",
