+++ src/AppBundle/BileMoAppBundle.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace AppBundle;
+
+use AppBundle\DependencyInjection\Compiler\ExceptionNormalizerPass;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\HttpKernel\Bundle\Bundle;
+
+class BileMoAppBundle extends Bundle
+{
+    public function build(ContainerBuilder $container)
+    {
+        $container->addCompilerPass(new ExceptionNormalizerPass());
+    }
+}
