+++ src/AppBundle/Exception/ResourceValidationException.php
@@ -1,0 +1,6 @@
+<?php
+namespace AppBundle\Exception;
+
+class ResourceValidationException extends \Exception
+{
+}
