+++ src/AppBundle/Normalizer/NormalizerInterface.php
@@ -1,0 +1,10 @@
+<?php
+namespace AppBundle\Normalizer;
+
+interface NormalizerInterface
+{
+    public function normalize(\Exception $exception);
+
+    public function supports(\Exception $exception);
+
+}
