+++ src/AppBundle/Normalizer/NotFoundHttpExceptionNormalizer.php
@@ -1,0 +1,21 @@
+<?php
+namespace AppBundle\Normalizer;
+
+use Symfony\Component\HttpFoundation\Response;
+
+class NotFoundHttpExceptionNormalizer extends AbstractNormalizer
+{
+
+    public function normalize(\Exception $exception)
+    {
+        $result['code'] = Response::HTTP_NOT_FOUND;
+
+        $result['body'] = [
+            'code' => Response::HTTP_NOT_FOUND,
+            'message' => 'The resource you requested cannot be found.'
+        ];
+
+        return $result;
+
+    }
+}
