--- src/BileMo/AppBundle/Entity/OAuth/AccessToken.php
@@ -1,30 +1,0 @@
-<?php
-namespace BileMo\AppBundle\Entity\OAuth;
-
-use FOS\OAuthServerBundle\Entity\AccessToken as BaseAccessToken;
-use Doctrine\ORM\Mapping as ORM;
-
-/**
- * @ORM\Entity
- */
-class AccessToken extends BaseAccessToken
-{
-    /**
-     * @ORM\Id
-     * @ORM\Column(type="integer")
-     * @ORM\GeneratedValue(strategy="AUTO")
-     */
-    protected $id;
-
-    /**
-     * @ORM\ManyToOne(targetEntity="Client")
-     * @ORM\JoinColumn(nullable=false)
-     */
-    protected $client;
-
-    /**
-     * @ORM\ManyToOne(targetEntity="BileMo\AppBundle\Entity\User")
-     * @ORM\JoinColumn(name="user_id", referencedColumnName="id", onDelete="CASCADE")
-     */
-    protected $user;
-}
