--- src/BileMo/AppBundle/Entity/OAuth/Client.php
@@ -1,24 +1,0 @@
-<?php
-namespace BileMo\AppBundle\Entity\OAuth;
-
-use FOS\OAuthServerBundle\Entity\Client as BaseClient;
-use Doctrine\ORM\Mapping as ORM;
-
-/**
- * @ORM\Entity
- */
-class Client extends BaseClient
-{
-    /**
-     * @ORM\Id
-     * @ORM\Column(type="integer")
-     * @ORM\GeneratedValue(strategy="AUTO")
-     */
-    protected $id;
-
-    public function __construct()
-    {
-        parent::__construct();
-        // your own logic
-    }
-}
