--- src/BileMo/AppBundle/Entity/User.php
@@ -1,404 +1,0 @@
-<?php
-
-namespace BileMo\AppBundle\Entity;
-
-use Doctrine\ORM\Mapping as ORM;
-use Gedmo\Mapping\Annotation as Gedmo;
-use Hateoas\Configuration\Annotation as Hateoas;
-use JMS\Serializer\Annotation\ExclusionPolicy;
-use JMS\Serializer\Annotation\Expose;
-use Symfony\Bridge\Doctrine\Validator\Constraints\UniqueEntity;
-use Symfony\Component\Security\Core\User\UserInterface;
-use Symfony\Component\Validator\Constraints as Assert;
-use Symfony\Component\Security\Core\Validator\Constraints as SecurityAssert;
-
-
-/**
- * @ORM\Table(name="users")
- * @ORM\Entity(repositoryClass="BileMo\AppBundle\Repository\UserRepository")
- * @UniqueEntity(fields="email", message="Email already taken, please choose another one.")
- * @UniqueEntity(fields="username", message="Username already taken, please choose another one.")
- * @ExclusionPolicy("all")
- * @Hateoas\Relation(
- *      "self",
- *      href = @Hateoas\Route(
- *          "show_user_details",
- *          parameters = { "id" = "expr(object.getId())" },
- *          absolute = true
- *      )
- * )
- * @Hateoas\Relation(
- *      "create",
- *      href = @Hateoas\Route(
- *          "create_user",
- *          absolute = true
- *      )
- * )
- * @Hateoas\Relation(
- *      "delete",
- *      href = @Hateoas\Route(
- *          "delete_user",
- *          parameters = { "id" = "expr(object.getId())" },
- *          absolute = true
- *      )
- * )
-
- */
-class User implements UserInterface, \Serializable
-{
-    /**
-     * @ORM\Column(type="integer")
-     * @ORM\Id
-     * @ORM\GeneratedValue(strategy="AUTO")
-     */
-    private $id;
-
-    /**
-     * @ORM\Column(type="string", length=25, unique=true)
-     * @Assert\NotBlank()
-     * @Assert\Regex(
-     *     pattern="/^[A-Za-z0-9-_]+$/",
-     *     message="Username can only contain letters (a-z), numbers or following characters : -_"
-     * )
-     * @Assert\Length(
-     *     min=3,
-     *     minMessage="Username must contain at least 3 characters.",
-     *     max=25,
-     *     maxMessage="Username must not contain more than 25 characters."
-     * )
-     * @Expose
-     *
-     */
-    private $username;
-
-    /**
-     * @Assert\NotBlank(
-     *     message="User's password cannot be empty."
-     * )
-     * @Assert\Length(
-     *     min=8,
-     *     minMessage="User's password must contain at least 8 characters",
-     *     max=4096,
-     * )
-     */
-    private $plainPassword;
-
-    /**
-     * @ORM\Column(type="string", length=100)
-     * @Assert\NotBlank(
-     *     message="User's first name cannot be empty."
-     * )
-     * @Assert\Regex(
-     *     pattern="/^[A-Za-z0-9-_\. ]+$/",
-     *     message="User's first name can only contain letters (a-z), numbers or following characters : -_."
-     * )
-     * @Assert\Length(
-     *     min=2,
-     *     minMessage="User's firstName must contain at least 2 characters.",
-     *     max=100,
-     *     maxMessage="User's firstName must not contain more than 100 characters."
-     * )
-     *
-     * @Expose
-     */
-    private $firstName;
-
-    /**
-     * @ORM\Column(type="string", length=100)
-     * @Assert\NotBlank(
-     *     message="User's last name cannot be empty."
-     * )
-     * @Assert\Regex(
-     *     pattern="/^[A-Za-z0-9-_\. ]+$/",
-     *     message="User's last name can only contain letters (a-z), numbers or following characters : -_."
-     * )
-     * @Assert\Length(
-     *     min=2,
-     *     minMessage="User's last name must contain at least 2 characters.",
-     *     max=100,
-     *     maxMessage="User's last name must not contain more than 100 characters."
-     * )
-     *
-     * @Expose
-     */
-    private $lastName;
-
-    /**
-     * @ORM\Column(type="string", length=64)
-     */
-    private $password;
-
-    /**
-     * @ORM\Column(type="datetime", nullable=false)
-     *
-     * @Expose
-     */
-    private $createdAt;
-
-    /**
-     * @ORM\Column(type="datetime", nullable=true)
-     *
-     * @Gedmo\Timestampable(on="update")
-     */
-    private $updatedAt;
-
-    /**
-     * @ORM\Column(type="string", length=254, unique=true)
-     * @Assert\NotBlank()
-     * @Assert\Email()
-     *
-     * @Expose
-     */
-    private $email;
-
-    /**
-     * @ORM\Column(type="array")
-     */
-    private $roles;
-
-
-    public function __construct()
-    {
-        $this->createdAt = new \DateTime();
-        $this->roles = array('ROLE_USER');
-    }
-
-    public function getId()
-    {
-        return $this->id;
-    }
-
-    public function getEmail()
-    {
-        return $this->email;
-    }
-
-    public function getUsername()
-    {
-        return $this->username;
-    }
-
-    public function getSalt()
-    {
-        return null;
-    }
-
-    public function getPassword()
-    {
-        return $this->password;
-    }
-
-    public function getRoles()
-    {
-        return $this->roles;
-    }
-
-    public function eraseCredentials()
-    {
-        $this->plainPassword = null;
-    }
-
-    /** @see \Serializable::serialize() */
-    public function serialize()
-    {
-        return serialize(array(
-            $this->id,
-            $this->username,
-            $this->password,
-        ));
-    }
-
-    /** @see \Serializable::unserialize() */
-    public function unserialize($serialized)
-    {
-        list (
-            $this->id,
-            $this->username,
-            $this->password
-        ) = unserialize($serialized, array('allowed_classes' => false));
-    }
-
-    /**
-     * Set username
-     *
-     * @param string $username
-     *
-     * @return User
-     */
-    public function setUsername($username)
-    {
-        $this->username = $username;
-
-        return $this;
-    }
-
-    /**
-     * Set password
-     *
-     * @param string $password
-     *
-     * @return User
-     */
-    public function setPassword($password)
-    {
-        $this->password = $password;
-
-        return $this;
-    }
-
-    /**
-     * Set email
-     *
-     * @param string $email
-     *
-     * @return User
-     */
-    public function setEmail($email)
-    {
-        $this->email = $email;
-
-        return $this;
-    }
-
-    /**
-     * Set roles
-     *
-     * @param array $roles
-     *
-     * @return User
-     */
-    public function setRoles($roles)
-    {
-        $this->roles = $roles;
-
-        return $this;
-    }
-
-    /**
-     * Get plainPassword
-     *
-     * @return string
-     */
-    public function getPlainPassword()
-    {
-        return $this->plainPassword;
-    }
-
-    /**
-     * Set plainPassword
-     *
-     * @param string $password
-     *
-     * @return User
-     */
-    public function setPlainPassword($plainPassword)
-    {
-        $this->plainPassword = $plainPassword;
-
-        /*
-         * We need to add this line because Doctrine listeners are not called
-         * if Doctrine thinks that an object has not been updated.
-         * A listener will be called in order to encode the password
-         */
-        $this->password = null;
-
-        return $this;
-    }
-
-    /**
-     * Set createdAt
-     *
-     * @param \DateTime $createdAt
-     *
-     * @return User
-     */
-    public function setCreatedAt($createdAt)
-    {
-        $this->createdAt = $createdAt;
-
-        return $this;
-    }
-
-    /**
-     * Get createdAt
-     *
-     * @return \DateTime
-     */
-    public function getCreatedAt()
-    {
-        return $this->createdAt;
-    }
-
-
-    /**
-     * Set firstName
-     *
-     * @param string $firstName
-     *
-     * @return User
-     */
-    public function setFirstName($firstName)
-    {
-        $this->firstName = $firstName;
-
-        return $this;
-    }
-
-    /**
-     * Get firstName
-     *
-     * @return string
-     */
-    public function getFirstName()
-    {
-        return $this->firstName;
-    }
-
-    /**
-     * Set lastName
-     *
-     * @param string $lastName
-     *
-     * @return User
-     */
-    public function setLastName($lastName)
-    {
-        $this->lastName = $lastName;
-
-        return $this;
-    }
-
-    /**
-     * Get lastName
-     *
-     * @return string
-     */
-    public function getLastName()
-    {
-        return $this->lastName;
-    }
-
-    /**
-     * Set updatedAt
-     *
-     * @param \DateTime $updatedAt
-     *
-     * @return User
-     */
-    public function setUpdatedAt($updatedAt)
-    {
-        $this->updatedAt = $updatedAt;
-
-        return $this;
-    }
-
-    /**
-     * Get updatedAt
-     *
-     * @return \DateTime
-     */
-    public function getUpdatedAt()
-    {
-        return $this->updatedAt;
-    }
-}
