--- src/BileMo/AppBundle/Exception/ResourceValidationException.php
@@ -1,6 +1,0 @@
-<?php
-namespace BileMo\AppBundle\Exception;
-
-class ResourceValidationException extends \Exception
-{
-}
