--- src/BileMo/AppBundle/Normalizer/AbstractNormalizer.php
@@ -1,18 +1,0 @@
-<?php
-
-namespace BileMo\AppBundle\Normalizer;
-
-abstract class AbstractNormalizer implements NormalizerInterface
-{
-    protected $exceptionTypes;
-
-    public function __construct(array $exceptionTypes)
-    {
-        $this->exceptionTypes = $exceptionTypes;
-    }
-
-    public function supports(\Exception $exception)
-    {
-        return in_array(get_class($exception), $this->exceptionTypes);
-    }
-}
