--- src/BileMo/AppBundle/Normalizer/NormalizerInterface.php
@@ -1,10 +1,0 @@
-<?php
-namespace BileMo\AppBundle\Normalizer;
-
-interface NormalizerInterface
-{
-    public function normalize(\Exception $exception);
-
-    public function supports(\Exception $exception);
-
-}
