--- src/BileMo/AppBundle/Repository/MobilePhoneRepository.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace BileMo\AppBundle\Repository;
-
-/**
- * MobilePhoneRepository
- */
-class MobilePhoneRepository extends AbstractRepository
-{
-    public function findAllPaginated($limit = 20, $offset = 0, $attributeToOrderBy = '', $order = 'asc', $relatedEntityClass = 'BileMo\AppBundle\Entity\MobilePhone' )
-    {
-        // Last attribute is Entity related Class, so that the method can check if the entity class has attribute to order results by.
-        return parent::findAllPaginated($limit, $offset, $attributeToOrderBy, $order, $relatedEntityClass);
-    }
-
-}
