--- src/BileMo/AppBundle/Security/UserProvider.php
@@ -1,60 +1,0 @@
-<?php
-namespace BileMo\AppBundle\Security;
-
-use Symfony\Component\Security\Core\User\UserInterface;
-use Symfony\Component\Security\Core\User\UserProviderInterface;
-use Symfony\Component\Security\Core\Exception\UsernameNotFoundException;
-use Symfony\Component\Security\Core\Exception\UnsupportedUserException;
-use Doctrine\Common\Persistence\ObjectRepository;
-use Doctrine\ORM\NoResultException;
-
-class UserProvider implements UserProviderInterface
-{
-    protected $userRepository;
-
-    public function __construct(ObjectRepository $userRepository){
-        $this->userRepository = $userRepository;
-    }
-
-    public function loadUserByUsername($username)
-    {
-        $q = $this->userRepository
-            ->createQueryBuilder('u')
-            ->where('u.username = :username')
-            ->setParameter('username', $username)
-            ->getQuery();
-
-        try {
-            $user = $q->getSingleResult();
-        } catch (NoResultException $e) {
-            $message = sprintf(
-                'Unable to find an active user identified by "%s". Please verify username.',
-                $username
-            );
-            throw new UsernameNotFoundException($message, 0, $e);
-        }
-
-        return $user;
-    }
-
-    public function refreshUser(UserInterface $user)
-    {
-        $class = get_class($user);
-        if (!$this->supportsClass($class)) {
-            throw new UnsupportedUserException(
-                sprintf(
-                    'Instances of "%s" are not supported.',
-                    $class
-                )
-            );
-        }
-
-        return $this->userRepository->find($user->getId());
-    }
-
-    public function supportsClass($class)
-    {
-        return $this->userRepository->getClassName() === $class
-            || is_subclass_of($class, $this->userRepository->getClassName());
-    }
-}
